pyproject.toml
[project]
name = "speech-in-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.1",
    "piper-tts>=1.3.0",
]
